feat: add suggestion mode and adjust pause detection
- Introduce dedicated SuggestionTaskQueue for handling suggestion generation
- Add Setting.SuggestionMode to toggle between translation and suggestion flows
- Lower PAUSE_DETECTION_MS from 3000 ms to 1000 ms for faster caption updates
- Refactor main loop into ProcessSuggestionMode/ProcessTranslationMode helpers
- Increase main window height in settings to accommodate new UI elements

+++ src/models/TranslateAPIConfig.cs
@@ -74,7 +74,7 @@ namespace LiveCaptionsTranslator.models
             public Message message { get; set; }
             public bool done { get; set; }
             public long total_duration { get; set; }
-            public int load_duration { get; set; }
+            public long load_duration { get; set; }
             public int prompt_eval_count { get; set; }
             public long prompt_eval_duration { get; set; }
             public int eval_count { get; set; }
